back-end/requirements.txt
fastapi
uvicorn
llama-cpp-python
PyMuPDF
python-multipart